refactor(server): rename internal/reearth-accounts-admin to internal/admin
The reearth-accounts- prefix is redundant inside the module's own
internal/ tree (the module is already reearth-accounts). Shorten the
admin bounded context to internal/admin so it can sit symmetrically with
the future internal/accounts (user API). The cmd/reearth-accounts-admin
binary keeps its descriptive name (it becomes a binary on PATH).

Updates all import paths plus the wire-admin/swag-admin Makefile targets;
regenerated wire_gen.go and docs/admin.
