+++ server/internal/admin/di/wire_gateway.go
