+++ server/internal/admin/di/wire_usecase.go
@@ -5,8 +5,8 @@ package di
 
 import (
 	"github.com/goforj/wire"
-	"github.com/reearth/reearth-accounts/server/internal/reearth-accounts-admin/usecase/authz"
-	"github.com/reearth/reearth-accounts/server/internal/reearth-accounts-admin/usecase/useruc"
+	"github.com/reearth/reearth-accounts/server/internal/admin/usecase/authz"
+	"github.com/reearth/reearth-accounts/server/internal/admin/usecase/useruc"
 )
 
 // usecaseWire provides the admin authorization checker and the usecases
