+++ server/internal/admin/usecase/authz/checker.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/cerbos/cerbos-sdk-go/cerbos"
 	effectv1 "github.com/cerbos/cerbos/api/genpb/cerbos/effect/v1"
-	adminrbac "github.com/reearth/reearth-accounts/server/internal/reearth-accounts-admin/rbac"
+	adminrbac "github.com/reearth/reearth-accounts/server/internal/admin/rbac"
 	"github.com/reearth/reearth-accounts/server/pkg/permittable"
 	"github.com/reearth/reearth-accounts/server/pkg/role"
 	"github.com/reearth/reearth-accounts/server/pkg/user"
